requirements.txt
fastapi
uvicorn
motor
bcrypt
pyjwt
python-dotenv
pydantic
